backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
requests
pydantic
supabase
bcrypt
python-dotenv
langchain-openai
langchain-community
openai
